transformation/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
pymongo>=4.3.0
python-dotenv>=0.21.0
pyspark>=3.5.0
google-generativeai>=0.8.3